Refactor: #31 - 불필요한 코드 리팩토링
- 인덴트 수정
- 로깅 코드 삭제
