+++ src/main/java/com/fourfourfour/damoa/common/entity/BaseLastModifiedEntity.java
@@ -22,5 +22,4 @@ public abstract class BaseLastModifiedEntity extends BaseCreatedEntity {
     @LastModifiedBy
     @Column(nullable = false)
     private String lastModifiedBy;
-
 }
